# core-lending-mortgages-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-mortgages
  description: Mortgages Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/mortgages
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/mortgage-properties/{propertyId}:
    get:
      tags:
        - MortgageProperty
      summary: Get a property record by ID
      operationId: getById
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MortgagePropertyDTO'
    put:
      tags:
        - MortgageProperty
      summary: Update a property record
      operationId: update
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MortgagePropertyDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MortgagePropertyDTO'
    delete:
      tags:
        - MortgageProperty
      summary: Delete a property record
      operationId: delete
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/mortgage-properties:
    get:
      tags:
        - MortgageProperty
      summary: List or search mortgage properties
      operationId: findAll
      parameters:
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.propertyType
          in: query
          description: Exact filter for property type
          required: false
          schema:
            type: string
            enum:
              - RESIDENTIAL
              - COMMERCIAL
              - LAND
              - INDUSTRIAL
        - name: filters.propertyStatus
          in: query
          description: Exact filter for property status
          required: false
          schema:
            type: string
            enum:
              - NEW
              - USED
              - OFF_PLAN
              - UNDER_CONSTRUCTION
        - name: filters.propertyUse
          in: query
          description: Exact filter for property use
          required: false
          schema:
            type: string
            enum:
              - PRIMARY_RESIDENCE
              - SECOND_HOME
              - INVESTMENT
        - name: filters.addressLine1
          in: query
          description: Exact filter for address line1
          required: false
          schema:
            type: string
        - name: filters.addressLine2
          in: query
          description: Exact filter for address line2
          required: false
          schema:
            type: string
        - name: filters.city
          in: query
          description: Exact filter for city
          required: false
          schema:
            type: string
        - name: filters.state
          in: query
          description: Exact filter for state
          required: false
          schema:
            type: string
        - name: filters.postalCode
          in: query
          description: Exact filter for postal code
          required: false
          schema:
            type: string
        - name: filters.countryCode
          in: query
          description: Exact filter for country code
          required: false
          schema:
            type: string
        - name: filters.landArea
          in: query
          description: Exact filter for land area
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[landArea].from
          in: query
          description: Filter land area from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[landArea].to
          in: query
          description: Filter land area to value
          required: false
          schema:
            type: string
        - name: filters.builtArea
          in: query
          description: Exact filter for built area
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[builtArea].from
          in: query
          description: Filter built area from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[builtArea].to
          in: query
          description: Filter built area to value
          required: false
          schema:
            type: string
        - name: filters.constructionYear
          in: query
          description: Exact filter for construction year
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[constructionYear].from
          in: query
          description: Filter construction year from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[constructionYear].to
          in: query
          description: Filter construction year to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.renovationYear
          in: query
          description: Exact filter for renovation year
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[renovationYear].from
          in: query
          description: Filter renovation year from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[renovationYear].to
          in: query
          description: Filter renovation year to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.titleNumber
          in: query
          description: Exact filter for title number
          required: false
          schema:
            type: string
        - name: filters.cadastralReference
          in: query
          description: Exact filter for cadastral reference
          required: false
          schema:
            type: string
        - name: filters.legalDescription
          in: query
          description: Exact filter for legal description
          required: false
          schema:
            type: string
        - name: filters.totalRooms
          in: query
          description: Exact filter for total rooms
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[totalRooms].from
          in: query
          description: Filter total rooms from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[totalRooms].to
          in: query
          description: Filter total rooms to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.totalBedrooms
          in: query
          description: Exact filter for total bedrooms
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[totalBedrooms].from
          in: query
          description: Filter total bedrooms from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[totalBedrooms].to
          in: query
          description: Filter total bedrooms to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.totalBathrooms
          in: query
          description: Exact filter for total bathrooms
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[totalBathrooms].from
          in: query
          description: Filter total bathrooms from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[totalBathrooms].to
          in: query
          description: Filter total bathrooms to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.hasParking
          in: query
          description: Exact filter for has parking
          required: false
          schema:
            type: boolean
        - name: filters.parkingSpaces
          in: query
          description: Exact filter for parking spaces
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[parkingSpaces].from
          in: query
          description: Filter parking spaces from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[parkingSpaces].to
          in: query
          description: Filter parking spaces to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.hasStorage
          in: query
          description: Exact filter for has storage
          required: false
          schema:
            type: boolean
        - name: filters.storageArea
          in: query
          description: Exact filter for storage area
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[storageArea].from
          in: query
          description: Filter storage area from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[storageArea].to
          in: query
          description: Filter storage area to value
          required: false
          schema:
            type: string
        - name: filters.hasElevator
          in: query
          description: Exact filter for has elevator
          required: false
          schema:
            type: boolean
        - name: filters.floorNumber
          in: query
          description: Exact filter for floor number
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[floorNumber].from
          in: query
          description: Filter floor number from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[floorNumber].to
          in: query
          description: Filter floor number to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.energyRating
          in: query
          description: Exact filter for energy rating
          required: false
          schema:
            type: string
            enum:
              - A
              - B
              - C
              - D
              - E
              - F
              - G
              - UNKNOWN
        - name: filters.restrictions
          in: query
          description: Exact filter for restrictions
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseMortgagePropertyDTO'
    post:
      tags:
        - MortgageProperty
      summary: Create a new property record
      operationId: create
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MortgagePropertyDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MortgagePropertyDTO'
...
components:
  schemas:
    MortgagePropertyDTO:
      required:
        - addressLine1
        - city
        - countryCode
        - hasElevator
        - hasParking
        - hasStorage
        - postalCode
        - propertyStatus
        - propertyType
        - propertyUse
        - state
      type: object
      properties:
        propertyId:
          type: string
          format: uuid
          readOnly: true
        propertyType:
          type: string
          enum:
            - RESIDENTIAL
            - COMMERCIAL
            - LAND
            - INDUSTRIAL
        propertyStatus:
          type: string
          enum:
            - NEW
            - USED
            - OFF_PLAN
            - UNDER_CONSTRUCTION
        propertyUse:
          type: string
          enum:
            - PRIMARY_RESIDENCE
            - SECOND_HOME
            - INVESTMENT
        addressLine1:
          maxLength: 200
          minLength: 0
          type: string
        addressLine2:
          maxLength: 200
          minLength: 0
          type: string
        city:
          maxLength: 100
          minLength: 0
          type: string
        state:
          maxLength: 100
          minLength: 0
          type: string
        postalCode:
          maxLength: 20
          minLength: 0
          type: string
        countryCode:
          maxLength: 3
          minLength: 2
          type: string
        landArea:
          maximum: 999999.99
          exclusiveMaximum: false
          minimum: 0.01
          exclusiveMinimum: false
          type: number
        builtArea:
          maximum: 999999.99
          exclusiveMaximum: false
          minimum: 0.01
          exclusiveMinimum: false
          type: number
        constructionYear:
          maximum: 2100
          minimum: 1800
          type: integer
          format: int32
        renovationYear:
          maximum: 2100
          minimum: 1800
          type: integer
          format: int32
        titleNumber:
          maxLength: 100
          minLength: 0
          type: string
        cadastralReference:
          maxLength: 100
          minLength: 0
          type: string
        legalDescription:
          maxLength: 1000
          minLength: 0
          type: string
        totalRooms:
          maximum: 100
          minimum: 1
          type: integer
          format: int32
        totalBedrooms:
          maximum: 50
          minimum: 0
          type: integer
          format: int32
        totalBathrooms:
          maximum: 50
          minimum: 0
          type: integer
          format: int32
        hasParking:
          type: boolean
        parkingSpaces:
          maximum: 100
          minimum: 0
          type: integer
          format: int32
        hasStorage:
          type: boolean
        storageArea:
          maximum: 9999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        hasElevator:
          type: boolean
        floorNumber:
          maximum: 200
          minimum: -10
          type: integer
          format: int32
        energyRating:
          type: string
          enum:
            - A
            - B
            - C
            - D
            - E
            - F
            - G
            - UNKNOWN
        restrictions:
          maxLength: 2000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseMortgagePropertyDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/MortgagePropertyDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...